app/requirements.txt
APScheduler==3.10.4
backports.zoneinfo==0.2.1
Brotli==1.1.0
cachelib==0.6.0
certifi==2024.7.4
charset-normalizer==2.0.12
click==8.0.4
ConfigArgParse==1.7
dataclasses==0.8
Flask==2.0.3
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.1
Flask-SQLAlchemy==2.5.1
gevent==22.10.2
geventhttpclient==2.0.12
greenlet==2.0.2
idna==3.7
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
locust==2.8.6
MarkupSafe==2.0.1
msgpack==1.0.5
psutil==6.0.0
PyMySQL==1.0.2
python-dotenv
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==25.1.2
requests==2.27.1
roundrobin==0.0.4
six==1.16.0
SQLAlchemy==1.4.54
typing_extensions==4.1.1
tzdata==2024.1
tzlocal==4.2
urllib3==1.26.19
waitress==2.0.0
Werkzeug==2.0.3
zipp==3.6.0
zope.event==4.6
zope.interface==5.5.2
